requirements.txt
elasticsearch
pyyaml
requests